Refactoring codebase to accommodate the connection object logic on each individual module.
This change was required to compensate for the lack of a main zowe class and,
 to avoid the need for the user to import the ApiConnection class in order to instantiate a Zowe module class
