+++ src/zos_jobs/zowe/zos_jobs_for_zowe_sdk/jobs.py
@@ -9,7 +9,7 @@ SPDX-License-Identifier: EPL-2.0
 
 Copyright Contributors to the Zowe Project.
 """
-from zowe.core import SdkApi
+from zowe.core_for_zowe_sdk import SdkApi
 import os
