+++ src/zos_tso/zowe/zos_tso_for_zowe_sdk/tso.py
@@ -10,7 +10,7 @@ SPDX-License-Identifier: EPL-2.0
 Copyright Contributors to the Zowe Project.
 """
 
-from zowe.core import SdkApi
+from zowe.core_for_zowe_sdk import SdkApi
 
 
 class Tso(SdkApi):
